+++ packages/client/src/StreamVideoClient.ts
@@ -48,8 +48,6 @@ export class StreamVideoClient {
    */
   constructor(apiKey: string, opts?: StreamClientOptions) {
     this.streamClient = new StreamClient(apiKey, {
-      // FIXME: OL: fix SSR.
-      browser: true,
       persistUserOnConnectionFailure: true,
       ...opts,
     });
@@ -370,4 +368,23 @@ export class StreamVideoClient {
       ...(userID ? { user_id: userID } : {}),
     });
   };
+
+  /**
+   * createToken - Creates a token to authenticate this user. This function is used server side.
+   * The resulting token should be passed to the client side when the users registers or logs in.
+   *
+   * @param {string} userID The User ID
+   * @param {number} [exp] The expiration time for the token expressed in the number of seconds since the epoch
+   * @param call_cids for anonymous tokens you have to provide the call cids the use can join
+   *
+   * @return {string} Returns a token
+   */
+  createToken(
+    userID: string,
+    exp?: number,
+    iat?: number,
+    call_cids?: string[],
+  ) {
+    return this.streamClient.createToken(userID, exp, iat, call_cids);
+  }
 }