+++ packages/react-native-dogfood/src/components/ActiveCallComponent.tsx
@@ -1,31 +1,17 @@
 import React, { useEffect } from 'react';
-import {
-  ActiveCall,
-  ActiveCallProps,
-  theme,
-  useCall,
-  useIncomingCalls,
-} from '@stream-io/video-react-native-sdk';
+import { ActiveCall, theme, useCall } from '@stream-io/video-react-native-sdk';
 import { ActivityIndicator, SafeAreaView, StyleSheet } from 'react-native';
-import { callkeepCallId$ } from '../../hooks/useCallkeepEffect';
+import { callkeepCallId$ } from '../hooks/useCallkeepEffect';
 import {
   startForegroundService,
   stopForegroundService,
-} from '../../modules/push/android';
-import { ParticipantListButtons } from '../../components/ParticipantListButtons';
+} from '../modules/push/android';
 
-type Mode = NonNullable<ActiveCallProps['mode']>;
-
-export const CallScreen = () => {
+export const ActiveCallComponent = () => {
   const call = useCall();
-  const [incomingCall] = useIncomingCalls();
-  const [selectedMode, setMode] = React.useState<Mode>('grid');
 
+  // effect to answer call when incoming call is received from callkeep
   useEffect(() => {
-    // effect to answer call when incoming call is received from callkeep
-    if (!incomingCall) {
-      return;
-    }
     const subscription = callkeepCallId$.subscribe((callkeepCallId) => {
       if (!callkeepCallId || !call) {
         return;
@@ -35,7 +21,7 @@ export const CallScreen = () => {
       callkeepCallId$.next(undefined); // remove the current call id to avoid rejoining when coming back to this screen
     });
     return () => subscription.unsubscribe();
-  }, [call, incomingCall]);
+  }, [call]);
 
   useEffect(() => {
     if (!call) {
@@ -52,8 +38,7 @@ export const CallScreen = () => {
   }
   return (
     <SafeAreaView style={styles.wrapper}>
-      <ParticipantListButtons selectedMode={selectedMode} setMode={setMode} />
-      <ActiveCall mode={selectedMode} />
+      <ActiveCall />
     </SafeAreaView>
   );
 };