+++ packages/react-native-dogfood/types.ts
@@ -132,14 +132,6 @@ export type MeetingStackParamList = {
   GuestMeetingScreen: GuestMeetingScreenParams;
 };
 
-export type CallStackParamList = {
-  JoinCallScreen: undefined;
-  CallScreen: undefined;
-  IncomingCallScreen: undefined;
-  OutgoingCallScreen: undefined;
-  CallParticipantsInfoScreen: undefined;
-};
-
 export type RootStackParamList = {
   Login: undefined;
   Meeting: undefined;