+++ packages/react-bindings/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.0.6](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.0.5...@stream-io/video-react-bindings-0.0.6) (2023-06-07)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.0.1`
 ### [0.0.5](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.0.4...@stream-io/video-react-bindings-0.0.5) (2023-06-06)
 
 ### Dependency Updates